requirements.txt
facenet-pytorch
fastapi
openai-clip
Pillow==11.2.1
pydantic
numpy
python-multipart
requests
scikit-learn
torch
torchaudio
torchvision
uvicorn[standard]
transformers
